requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile setup.py
#
filemagic==1.6
    # via helix
helix==1.1.1.post1
    # via blind-helix (setup.py)
lief==0.11.5
    # via blind-helix (setup.py)
